pyproject.toml
[tool.poetry]
name = "parsee-pdf-reader"
version = "0.1.0"
description = ""
authors = ["Parsee.ai <info@parsee.ai>"]
readme = "README.md"
homepage = "https://parsee.ai"
packages = [{include = "pdf_reader"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pillow = "^10.0.0"
numpy = "^1.25.1"
pdfminer-six = "^20221105"
pytesseract = "^0.3.10"
opencv-python = "^4.8.0.74"
pdf2image = "^1.16.3"
pypdf = "^3.16.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
